requirements.txt
numpy
torch
torchvision
open_clip_torch
pillow
tqdm
ftfy
regex
scipy
